requirements.txt
instabot
datetime
configparser
numpy
Pillow
requests
soupsieve
tqdm
urllib3
email
imaplib
smtplib
os